fix(loaders): isolate scheduler agent run identities
